requirements.txt
pytest==8.0.0
pytest-playwright==0.4.4
pytest-base-url==2.1.0
python-dotenv==1.0.1
supabase==2.0.0
allure-pytest==2.13.2
